Update README.md
This change is based on feedback I got from the hack-a-thon. Several users would have found it helpful to know about re-using an existing sender object before seeing the link to set up a new sender object. If this change is accepted, it needs to be added to similar steps in  other SDKs.